Migrate policingice app from TanStack Start to Next.js
Replace TanStack Start/Router with Next.js App Router, using React Server
Components for data fetching and Server Actions for mutations.

Key changes:
- Replace createServerFn() with "use server" actions in src/actions/
- Replace TanStack Router file-based routing with Next.js App Router
- RSC pages fetch data directly, pass to client components for interactivity
- Replace tanstackStartCookies with nextCookies for better-auth
- Replace Vite + Nitro build with Next.js + Turbopack
- Generate sitemap.xml and robots.txt via Next.js metadata API
- Auth API uses Next.js Route Handler with toNextJsHandler
