--- apps/policingice/.cta.json
@@ -1,16 +1,0 @@
-{
-  "projectName": "policingice",
-  "mode": "file-router",
-  "typescript": true,
-  "tailwind": true,
-  "packageManager": "npm",
-  "addOnOptions": {
-    "drizzle": {
-      "database": "sqlite"
-    }
-  },
-  "git": true,
-  "version": 1,
-  "framework": "react-cra",
-  "chosenAddOns": ["eslint", "nitro", "start", "drizzle"]
-}
