+++ apps/policingice/CLAUDE.md
@@ -5,9 +5,8 @@ This file provides guidance to Claude Code (claude.ai/code) when working with co
 ## Commands
 
 ```bash
-pnpm dev          # Dev server on port 3000
+pnpm dev          # Dev server on port 3000 (Next.js with Turbopack)
 pnpm build        # Production build
-pnpm test         # Run tests (vitest)
 pnpm check        # Format + lint fix
 pnpm db:push      # Push schema to Turso
 pnpm db:reset     # Clear content tables (incidents, videos, votes)
@@ -22,37 +21,43 @@ npx tsx scripts/create-admin.ts <email> <password> [name]
 
 ## Architecture
 
-**TanStack Start + Nitro full-stack app** - uses file-based routing with SSR capabilities.
+**Next.js App Router** - uses React Server Components and Server Actions.
 
 ### Key Layers
 
-- **Routes** (`src/routes/`): File-based routing via TanStack Router. Server functions defined with `createServerFn()`.
+- **App Router** (`src/app/`): Next.js file-based routing with RSC pages and Server Actions.
+- **Server Actions** (`src/actions/`): `incidents.ts` for public actions, `admin.ts` for admin actions. All marked with `"use server"`.
 - **Database** (`src/db/`): Drizzle ORM + Turso (SQLite). Schema in `schema.ts`, client in `index.ts`.
-- **Auth** (`src/lib/auth.ts`): better-auth with anonymous + email/password. Client in `auth-client.ts`.
+- **Auth** (`src/lib/auth.ts`): better-auth with anonymous + email/password + `nextCookies` plugin. Client in `auth-client.ts`.
+- **Components** (`src/components/`): Shared UI components, all client components with `"use client"`.
 
 ### Data Flow Pattern
 
-Routes use `createServerFn()` for server-side data fetching. These run on server during SSR and client-side after hydration:
+Pages are React Server Components that fetch data directly, then pass to client components for interactivity:
 
 ```tsx
-const getData = createServerFn({ method: "GET" })
-  .inputValidator((data) => data)
-  .handler(async ({ data }) => {
-    /* db queries */
-  });
+// src/app/page.tsx (RSC)
+export default async function Page() {
+  const data = await db.query.incidents.findMany(...);
+  return <ClientComponent data={data} />;
+}
 
-export const Route = createFileRoute("/path")({
-  loader: () => getData({ data: {} }),
-  component: Component,
-});
+// Mutations use Server Actions
+"use server";
+export async function createIncident(data) { ... }
 ```
 
 ### Route Structure
 
-- `/` - Public incident feed with infinite scroll + voting
-- `/incident/$id` - Single incident view
-- `/admin/*` - Protected admin routes (requires non-anonymous user)
-- `/api/auth/$` - better-auth API handler
+- `/` - Public incident feed with infinite scroll + voting (RSC + client feed)
+- `/incident/[id]` - Single incident view (RSC + client detail)
+- `/share` - Web Share API handler (RSC with redirect)
+- `/admin/login` - Admin login (client component)
+- `/admin/(dashboard)/*` - Protected admin routes (RSC layout with auth check)
+- `/api/auth/[...all]` - better-auth API handler (Route Handler)
+- `/api/rss` - RSS feed (Route Handler)
+- `/sitemap.ts` - Dynamic sitemap
+- `/robots.ts` - Robots.txt
 
 ### Schema
 
@@ -64,7 +69,7 @@ Core tables: `incidents` (with videos relation), `videos`, `votes`. Auth tables:
 
 ## Environment Variables
 
-Required in `.env.local`:
+Required in `.env`:
 
 - `TURSO_DATABASE_URL`
 - `TURSO_AUTH_TOKEN`
