--- apps/policingice/app.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from "@tanstack/react-start/config";
-
-export default defineConfig({
-  server: {
-    routeRules: {
-      "/sitemap.xml": {
-        redirect: "/api/sitemap",
-      },
-    },
-  },
-});
