+++ apps/policingice/package.json
@@ -3,15 +3,15 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "build": "pnpm with-env vite build",
+    "build": "pnpm with-env next build",
     "check": "prettier --write . && eslint --fix",
     "db:push": "pnpm with-env drizzle-kit push",
     "db:reset": "pnpm with-env npx tsx scripts/db-reset.ts",
     "db:studio": "pnpm with-env drizzle-kit studio",
-    "dev": "pnpm with-env vite dev",
+    "dev": "pnpm with-env next dev --turbopack -p 3000",
     "format": "prettier --check . --ignore-path ../../.gitignore",
     "lint": "eslint",
-    "preview": "pnpm with-env vite preview",
+    "start": "pnpm with-env next start",
     "typecheck": "tsc --noEmit",
     "with-env": "dotenv -e ../../.env --"
   },
@@ -19,13 +19,6 @@
     "@ai-sdk/xai": "^3.0.57",
     "@base-ui/react": "catalog:",
     "@libsql/client": "^0.17.0",
-    "@tailwindcss/vite": "^4.2.0",
-    "@tanstack/react-devtools": "^0.9.6",
-    "@tanstack/react-router": "^1.161.3",
-    "@tanstack/react-router-devtools": "^1.161.3",
-    "@tanstack/react-router-ssr-query": "^1.161.3",
-    "@tanstack/react-start": "^1.161.3",
-    "@tanstack/router-plugin": "^1.161.3",
     "@vercel/analytics": "catalog:",
     "ai": "^6.0.97",
     "better-auth": "^1.4.18",
@@ -33,28 +26,26 @@
     "drizzle-orm": "^0.45.1",
     "embla-carousel-react": "^8.6.0",
     "lucide-react": "catalog:",
-    "nitro": "npm:nitro-nightly@latest",
+    "next": "catalog:",
     "react": "catalog:",
     "react-dom": "catalog:",
     "react-tweet": "^3.3.0",
     "tailwindcss": "catalog:",
-    "vite-tsconfig-paths": "^6.1.1",
     "zod": "^4.3.6"
   },
   "devDependencies": {
+    "@kyh/eslint-config": "workspace:^",
     "@kyh/prettier-config": "workspace:^",
-    "@tanstack/devtools-vite": "^0.4.1",
-    "@tanstack/eslint-config": "^0.3.4",
+    "@kyh/tsconfig": "workspace:^",
+    "@tailwindcss/postcss": "^4.2.0",
     "@types/node": "catalog:",
     "@types/react": "catalog:",
     "@types/react-dom": "catalog:",
-    "@vitejs/plugin-react": "catalog:",
     "eslint": "catalog:",
+    "postcss": "^8.5.6",
     "prettier": "catalog:",
     "tsx": "catalog:",
-    "typescript": "catalog:",
-    "vite": "catalog:",
-    "web-vitals": "^5.1.0"
+    "typescript": "catalog:"
   },
   "prettier": "@kyh/prettier-config"
 }
