+++ apps/policingice/src/app/admin/(dashboard)/incidents/page.tsx
@@ -1,0 +1,10 @@
+import {
+  getAllIncidents,
+} from "@/actions/admin";
+
+import { AdminIncidentsClient } from "./incidents-client";
+
+export default async function AdminIncidentsPage() {
+  const allIncidents = await getAllIncidents();
+  return <AdminIncidentsClient initialIncidents={allIncidents} />;
+}
