+++ apps/policingice/src/app/admin/(dashboard)/page.tsx
@@ -1,0 +1,5 @@
+import { redirect } from "next/navigation";
+
+export default function AdminPage() {
+  redirect("/admin/incidents");
+}
