+++ apps/policingice/src/app/api/auth/[...all]/route.ts
@@ -1,0 +1,5 @@
+import { toNextJsHandler } from "better-auth/next-js";
+
+import { auth } from "@/lib/auth";
+
+export const { GET, POST } = toNextJsHandler(auth);
