+++ apps/policingice/src/app/layout.tsx
@@ -1,0 +1,64 @@
+import type { Metadata } from "next";
+import { Analytics } from "@vercel/analytics/react";
+
+import { ToastProvider } from "@/components/Toast";
+
+import "./globals.css";
+
+const siteUrl = "https://policingice.com";
+const siteName = "Policing ICE";
+const siteDescription =
+  "Documenting and tracking incidents of ICE overreach across the United States. Community-driven accountability through video evidence.";
+
+export const metadata: Metadata = {
+  title: siteName,
+  description: siteDescription,
+  metadataBase: new URL(siteUrl),
+  openGraph: {
+    type: "website",
+    siteName,
+    title: siteName,
+    description: siteDescription,
+    url: siteUrl,
+    images: [{ url: "/og-image.png", width: 1200, height: 630 }],
+  },
+  twitter: {
+    card: "summary_large_image",
+    title: siteName,
+    description: siteDescription,
+    images: ["/og-image.png"],
+  },
+  icons: [
+    { rel: "icon", type: "image/svg+xml", url: "/favicon.svg" },
+    { rel: "icon", type: "image/png", sizes: "96x96", url: "/favicon-96x96.png" },
+    { rel: "icon", type: "image/x-icon", url: "/favicon.ico" },
+    { rel: "apple-touch-icon", sizes: "180x180", url: "/apple-touch-icon.png" },
+  ],
+  manifest: "/site.webmanifest",
+  other: {
+    "theme-color": "#ffffff",
+  },
+};
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en">
+      <body>
+        <ToastProvider>
+          <a
+            href="#main-content"
+            className="sr-only focus:not-sr-only focus:absolute focus:top-4 focus:left-4 focus:z-50 focus:rounded focus:bg-neutral-900 focus:px-4 focus:py-2 focus:text-white"
+          >
+            Skip to content
+          </a>
+          {children}
+        </ToastProvider>
+        <Analytics />
+      </body>
+    </html>
+  );
+}
