+++ apps/policingice/src/app/page.tsx
@@ -1,0 +1,23 @@
+import { getAdminUser } from "@/lib/admin-auth";
+import { getIncidents, getUserVotes } from "@/actions/incidents";
+
+import { IncidentFeed } from "./incident-feed";
+
+export default async function HomePage() {
+  const [{ incidents, nextOffset }, admin] = await Promise.all([
+    getIncidents({}),
+    getAdminUser(),
+  ]);
+  const userVotes = await getUserVotes({
+    incidentIds: incidents.map((i) => i.id),
+  });
+
+  return (
+    <IncidentFeed
+      initialIncidents={incidents}
+      initialNextOffset={nextOffset}
+      initialUserVotes={userVotes}
+      isAdmin={!!admin}
+    />
+  );
+}
