+++ apps/policingice/src/app/robots.ts
@@ -1,0 +1,14 @@
+import type { MetadataRoute } from "next";
+
+export default function robots(): MetadataRoute.Robots {
+  return {
+    rules: [
+      {
+        userAgent: "*",
+        allow: "/",
+        disallow: "/admin",
+      },
+    ],
+    sitemap: "https://policingice.com/sitemap.xml",
+  };
+}
