+++ apps/policingice/src/lib/admin-auth.ts
@@ -1,17 +1,14 @@
-import { createServerFn } from "@tanstack/react-start";
-import { getRequestHeaders } from "@tanstack/react-start/server";
+import { headers } from "next/headers";
 
 import { auth } from "@/lib/auth";
 
-export const getAdminUser = createServerFn({ method: "GET" }).handler(
-  async () => {
-    const headers = getRequestHeaders();
-    const session = await auth.api.getSession({ headers });
+export async function getAdminUser() {
+  const headersList = await headers();
+  const session = await auth.api.getSession({ headers: headersList });
 
-    if (!session?.user || session.user.isAnonymous) {
-      return null;
-    }
+  if (!session?.user || session.user.isAnonymous) {
+    return null;
+  }
 
-    return session.user;
-  },
-);
+  return session.user;
+}
