--- apps/policingice/src/routeTree.gen.ts
@@ -1,311 +1,0 @@
-/* eslint-disable */
-
-// @ts-nocheck
-
-// noinspection JSUnusedGlobalSymbols
-
-// This file was automatically generated by TanStack Router.
-// You should NOT make any changes in this file as it will be overwritten.
-// Additionally, you should also exclude this file from your linter and/or formatter to prevent it from being checked or modified.
-
-import { Route as rootRouteImport } from './routes/__root'
-import { Route as ShareRouteImport } from './routes/share'
-import { Route as IndexRouteImport } from './routes/index'
-import { Route as IncidentIdRouteImport } from './routes/incident.$id'
-import { Route as ApiSitemapRouteImport } from './routes/api/sitemap'
-import { Route as ApiRssRouteImport } from './routes/api/rss'
-import { Route as AdminLoginRouteImport } from './routes/admin.login'
-import { Route as AdminLayoutRouteImport } from './routes/admin/_layout'
-import { Route as AdminLayoutIndexRouteImport } from './routes/admin/_layout/index'
-import { Route as ApiAuthSplatRouteImport } from './routes/api/auth/$'
-import { Route as AdminLayoutRedditFeedRouteImport } from './routes/admin/_layout/reddit-feed'
-import { Route as AdminLayoutIncidentsRouteImport } from './routes/admin/_layout/incidents'
-import { Route as AdminLayoutCreateRouteImport } from './routes/admin/_layout/create'
-
-const ShareRoute = ShareRouteImport.update({
-  id: '/share',
-  path: '/share',
-  getParentRoute: () => rootRouteImport,
-} as any)
-const IndexRoute = IndexRouteImport.update({
-  id: '/',
-  path: '/',
-  getParentRoute: () => rootRouteImport,
-} as any)
-const IncidentIdRoute = IncidentIdRouteImport.update({
-  id: '/incident/$id',
-  path: '/incident/$id',
-  getParentRoute: () => rootRouteImport,
-} as any)
-const ApiSitemapRoute = ApiSitemapRouteImport.update({
-  id: '/api/sitemap',
-  path: '/api/sitemap',
-  getParentRoute: () => rootRouteImport,
-} as any)
-const ApiRssRoute = ApiRssRouteImport.update({
-  id: '/api/rss',
-  path: '/api/rss',
-  getParentRoute: () => rootRouteImport,
-} as any)
-const AdminLoginRoute = AdminLoginRouteImport.update({
-  id: '/admin/login',
-  path: '/admin/login',
-  getParentRoute: () => rootRouteImport,
-} as any)
-const AdminLayoutRoute = AdminLayoutRouteImport.update({
-  id: '/admin/_layout',
-  path: '/admin',
-  getParentRoute: () => rootRouteImport,
-} as any)
-const AdminLayoutIndexRoute = AdminLayoutIndexRouteImport.update({
-  id: '/',
-  path: '/',
-  getParentRoute: () => AdminLayoutRoute,
-} as any)
-const ApiAuthSplatRoute = ApiAuthSplatRouteImport.update({
-  id: '/api/auth/$',
-  path: '/api/auth/$',
-  getParentRoute: () => rootRouteImport,
-} as any)
-const AdminLayoutRedditFeedRoute = AdminLayoutRedditFeedRouteImport.update({
-  id: '/reddit-feed',
-  path: '/reddit-feed',
-  getParentRoute: () => AdminLayoutRoute,
-} as any)
-const AdminLayoutIncidentsRoute = AdminLayoutIncidentsRouteImport.update({
-  id: '/incidents',
-  path: '/incidents',
-  getParentRoute: () => AdminLayoutRoute,
-} as any)
-const AdminLayoutCreateRoute = AdminLayoutCreateRouteImport.update({
-  id: '/create',
-  path: '/create',
-  getParentRoute: () => AdminLayoutRoute,
-} as any)
-
-export interface FileRoutesByFullPath {
-  '/': typeof IndexRoute
-  '/share': typeof ShareRoute
-  '/admin': typeof AdminLayoutRouteWithChildren
-  '/admin/login': typeof AdminLoginRoute
-  '/api/rss': typeof ApiRssRoute
-  '/api/sitemap': typeof ApiSitemapRoute
-  '/incident/$id': typeof IncidentIdRoute
-  '/admin/create': typeof AdminLayoutCreateRoute
-  '/admin/incidents': typeof AdminLayoutIncidentsRoute
-  '/admin/reddit-feed': typeof AdminLayoutRedditFeedRoute
-  '/api/auth/$': typeof ApiAuthSplatRoute
-  '/admin/': typeof AdminLayoutIndexRoute
-}
-export interface FileRoutesByTo {
-  '/': typeof IndexRoute
-  '/share': typeof ShareRoute
-  '/admin/login': typeof AdminLoginRoute
-  '/api/rss': typeof ApiRssRoute
-  '/api/sitemap': typeof ApiSitemapRoute
-  '/incident/$id': typeof IncidentIdRoute
-  '/admin/create': typeof AdminLayoutCreateRoute
-  '/admin/incidents': typeof AdminLayoutIncidentsRoute
-  '/admin/reddit-feed': typeof AdminLayoutRedditFeedRoute
-  '/api/auth/$': typeof ApiAuthSplatRoute
-  '/admin': typeof AdminLayoutIndexRoute
-}
-export interface FileRoutesById {
-  __root__: typeof rootRouteImport
-  '/': typeof IndexRoute
-  '/share': typeof ShareRoute
-  '/admin/_layout': typeof AdminLayoutRouteWithChildren
-  '/admin/login': typeof AdminLoginRoute
-  '/api/rss': typeof ApiRssRoute
-  '/api/sitemap': typeof ApiSitemapRoute
-  '/incident/$id': typeof IncidentIdRoute
-  '/admin/_layout/create': typeof AdminLayoutCreateRoute
-  '/admin/_layout/incidents': typeof AdminLayoutIncidentsRoute
-  '/admin/_layout/reddit-feed': typeof AdminLayoutRedditFeedRoute
-  '/api/auth/$': typeof ApiAuthSplatRoute
-  '/admin/_layout/': typeof AdminLayoutIndexRoute
-}
-export interface FileRouteTypes {
-  fileRoutesByFullPath: FileRoutesByFullPath
-  fullPaths:
-    | '/'
-    | '/share'
-    | '/admin'
-    | '/admin/login'
-    | '/api/rss'
-    | '/api/sitemap'
-    | '/incident/$id'
-    | '/admin/create'
-    | '/admin/incidents'
-    | '/admin/reddit-feed'
-    | '/api/auth/$'
-    | '/admin/'
-  fileRoutesByTo: FileRoutesByTo
-  to:
-    | '/'
-    | '/share'
-    | '/admin/login'
-    | '/api/rss'
-    | '/api/sitemap'
-    | '/incident/$id'
-    | '/admin/create'
-    | '/admin/incidents'
-    | '/admin/reddit-feed'
-    | '/api/auth/$'
-    | '/admin'
-  id:
-    | '__root__'
-    | '/'
-    | '/share'
-    | '/admin/_layout'
-    | '/admin/login'
-    | '/api/rss'
-    | '/api/sitemap'
-    | '/incident/$id'
-    | '/admin/_layout/create'
-    | '/admin/_layout/incidents'
-    | '/admin/_layout/reddit-feed'
-    | '/api/auth/$'
-    | '/admin/_layout/'
-  fileRoutesById: FileRoutesById
-}
-export interface RootRouteChildren {
-  IndexRoute: typeof IndexRoute
-  ShareRoute: typeof ShareRoute
-  AdminLayoutRoute: typeof AdminLayoutRouteWithChildren
-  AdminLoginRoute: typeof AdminLoginRoute
-  ApiRssRoute: typeof ApiRssRoute
-  ApiSitemapRoute: typeof ApiSitemapRoute
-  IncidentIdRoute: typeof IncidentIdRoute
-  ApiAuthSplatRoute: typeof ApiAuthSplatRoute
-}
-
-declare module '@tanstack/react-router' {
-  interface FileRoutesByPath {
-    '/share': {
-      id: '/share'
-      path: '/share'
-      fullPath: '/share'
-      preLoaderRoute: typeof ShareRouteImport
-      parentRoute: typeof rootRouteImport
-    }
-    '/': {
-      id: '/'
-      path: '/'
-      fullPath: '/'
-      preLoaderRoute: typeof IndexRouteImport
-      parentRoute: typeof rootRouteImport
-    }
-    '/incident/$id': {
-      id: '/incident/$id'
-      path: '/incident/$id'
-      fullPath: '/incident/$id'
-      preLoaderRoute: typeof IncidentIdRouteImport
-      parentRoute: typeof rootRouteImport
-    }
-    '/api/sitemap': {
-      id: '/api/sitemap'
-      path: '/api/sitemap'
-      fullPath: '/api/sitemap'
-      preLoaderRoute: typeof ApiSitemapRouteImport
-      parentRoute: typeof rootRouteImport
-    }
-    '/api/rss': {
-      id: '/api/rss'
-      path: '/api/rss'
-      fullPath: '/api/rss'
-      preLoaderRoute: typeof ApiRssRouteImport
-      parentRoute: typeof rootRouteImport
-    }
-    '/admin/login': {
-      id: '/admin/login'
-      path: '/admin/login'
-      fullPath: '/admin/login'
-      preLoaderRoute: typeof AdminLoginRouteImport
-      parentRoute: typeof rootRouteImport
-    }
-    '/admin/_layout': {
-      id: '/admin/_layout'
-      path: '/admin'
-      fullPath: '/admin'
-      preLoaderRoute: typeof AdminLayoutRouteImport
-      parentRoute: typeof rootRouteImport
-    }
-    '/admin/_layout/': {
-      id: '/admin/_layout/'
-      path: '/'
-      fullPath: '/admin/'
-      preLoaderRoute: typeof AdminLayoutIndexRouteImport
-      parentRoute: typeof AdminLayoutRoute
-    }
-    '/api/auth/$': {
-      id: '/api/auth/$'
-      path: '/api/auth/$'
-      fullPath: '/api/auth/$'
-      preLoaderRoute: typeof ApiAuthSplatRouteImport
-      parentRoute: typeof rootRouteImport
-    }
-    '/admin/_layout/reddit-feed': {
-      id: '/admin/_layout/reddit-feed'
-      path: '/reddit-feed'
-      fullPath: '/admin/reddit-feed'
-      preLoaderRoute: typeof AdminLayoutRedditFeedRouteImport
-      parentRoute: typeof AdminLayoutRoute
-    }
-    '/admin/_layout/incidents': {
-      id: '/admin/_layout/incidents'
-      path: '/incidents'
-      fullPath: '/admin/incidents'
-      preLoaderRoute: typeof AdminLayoutIncidentsRouteImport
-      parentRoute: typeof AdminLayoutRoute
-    }
-    '/admin/_layout/create': {
-      id: '/admin/_layout/create'
-      path: '/create'
-      fullPath: '/admin/create'
-      preLoaderRoute: typeof AdminLayoutCreateRouteImport
-      parentRoute: typeof AdminLayoutRoute
-    }
-  }
-}
-
-interface AdminLayoutRouteChildren {
-  AdminLayoutCreateRoute: typeof AdminLayoutCreateRoute
-  AdminLayoutIncidentsRoute: typeof AdminLayoutIncidentsRoute
-  AdminLayoutRedditFeedRoute: typeof AdminLayoutRedditFeedRoute
-  AdminLayoutIndexRoute: typeof AdminLayoutIndexRoute
-}
-
-const AdminLayoutRouteChildren: AdminLayoutRouteChildren = {
-  AdminLayoutCreateRoute: AdminLayoutCreateRoute,
-  AdminLayoutIncidentsRoute: AdminLayoutIncidentsRoute,
-  AdminLayoutRedditFeedRoute: AdminLayoutRedditFeedRoute,
-  AdminLayoutIndexRoute: AdminLayoutIndexRoute,
-}
-
-const AdminLayoutRouteWithChildren = AdminLayoutRoute._addFileChildren(
-  AdminLayoutRouteChildren,
-)
-
-const rootRouteChildren: RootRouteChildren = {
-  IndexRoute: IndexRoute,
-  ShareRoute: ShareRoute,
-  AdminLayoutRoute: AdminLayoutRouteWithChildren,
-  AdminLoginRoute: AdminLoginRoute,
-  ApiRssRoute: ApiRssRoute,
-  ApiSitemapRoute: ApiSitemapRoute,
-  IncidentIdRoute: IncidentIdRoute,
-  ApiAuthSplatRoute: ApiAuthSplatRoute,
-}
-export const routeTree = rootRouteImport
-  ._addFileChildren(rootRouteChildren)
-  ._addFileTypes<FileRouteTypes>()
-
-import type { getRouter } from './router.tsx'
-import type { createStart } from '@tanstack/react-start'
-declare module '@tanstack/react-start' {
-  interface Register {
-    ssr: true
-    router: Awaited<ReturnType<typeof getRouter>>
-  }
-}
