--- apps/policingice/src/router.tsx
@@ -1,17 +1,0 @@
-import { createRouter } from "@tanstack/react-router";
-
-// Import the generated route tree
-import { routeTree } from "./routeTree.gen";
-
-// Create a new router instance
-export const getRouter = () => {
-  const router = createRouter({
-    routeTree,
-    context: {},
-
-    scrollRestoration: true,
-    defaultPreloadStaleTime: 0,
-  });
-
-  return router;
-};
