--- apps/policingice/src/routes/admin/_layout/index.tsx
@@ -1,7 +1,0 @@
-import { createFileRoute, redirect } from "@tanstack/react-router";
-
-export const Route = createFileRoute("/admin/_layout/")({
-  beforeLoad: () => {
-    throw redirect({ to: "/admin/incidents" });
-  },
-});
