--- apps/policingice/src/routes/api/auth/$.ts
@@ -1,16 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-
-import { auth } from "@/lib/auth";
-
-export const Route = createFileRoute("/api/auth/$")({
-  server: {
-    handlers: {
-      GET: ({ request }) => {
-        return auth.handler(request);
-      },
-      POST: ({ request }) => {
-        return auth.handler(request);
-      },
-    },
-  },
-});
